+++ flows/engine/services.go
@@ -15,9 +15,6 @@ type WebhookServiceFactory func(flows.SessionAssets) (flows.WebhookService, erro
 // ClassificationServiceFactory resolves a session and classifier to an NLU service
 type ClassificationServiceFactory func(*flows.Classifier) (flows.ClassificationService, error)
 
-// TicketServiceFactory resolves a session to a ticket service
-type TicketServiceFactory func(*flows.Ticketer) (flows.TicketService, error)
-
 // AirtimeServiceFactory resolves a session to an airtime service
 type AirtimeServiceFactory func(flows.SessionAssets) (flows.AirtimeService, error)
 
@@ -25,7 +22,6 @@ type services struct {
 	email          EmailServiceFactory
 	webhook        WebhookServiceFactory
 	classification ClassificationServiceFactory
-	ticket         TicketServiceFactory
 	airtime        AirtimeServiceFactory
 }
 
@@ -39,9 +35,6 @@ func newEmptyServices() *services {
 		},
 		classification: func(*flows.Classifier) (flows.ClassificationService, error) {
 			return nil, errors.New("no classification service factory configured")
-		},
-		ticket: func(*flows.Ticketer) (flows.TicketService, error) {
-			return nil, errors.New("no ticket service factory configured")
 		},
 		airtime: func(flows.SessionAssets) (flows.AirtimeService, error) {
 			return nil, errors.New("no airtime service factory configured")
@@ -59,10 +52,6 @@ func (s *services) Webhook(sa flows.SessionAssets) (flows.WebhookService, error)
 
 func (s *services) Classification(classifier *flows.Classifier) (flows.ClassificationService, error) {
 	return s.classification(classifier)
-}
-
-func (s *services) Ticket(ticketer *flows.Ticketer) (flows.TicketService, error) {
-	return s.ticket(ticketer)
 }
 
 func (s *services) Airtime(sa flows.SessionAssets) (flows.AirtimeService, error) {