+++ flows/tickets.go
@@ -14,38 +14,31 @@ type TicketUUID uuids.UUID
 
 // Ticket is a ticket in a ticketing system
 type Ticket struct {
-	uuid       TicketUUID
-	ticketer   *Ticketer
-	topic      *Topic
-	body       string
-	externalID string
-	assignee   *User
+	uuid     TicketUUID
+	topic    *Topic
+	body     string
+	assignee *User
 }
 
 // NewTicket creates a new ticket
-func NewTicket(uuid TicketUUID, ticketer *Ticketer, topic *Topic, body, externalID string, assignee *User) *Ticket {
+func NewTicket(uuid TicketUUID, topic *Topic, body string, assignee *User) *Ticket {
 	return &Ticket{
-		uuid:       uuid,
-		ticketer:   ticketer,
-		topic:      topic,
-		body:       body,
-		externalID: externalID,
-		assignee:   assignee,
+		uuid:     uuid,
+		topic:    topic,
+		body:     body,
+		assignee: assignee,
 	}
 }
 
 // OpenTicket creates a new ticket. Used by ticketing services to open a new ticket.
-func OpenTicket(ticketer *Ticketer, topic *Topic, body string, assignee *User) *Ticket {
-	return NewTicket(TicketUUID(uuids.New()), ticketer, topic, body, "", assignee)
+func OpenTicket(topic *Topic, body string, assignee *User) *Ticket {
+	return NewTicket(TicketUUID(uuids.New()), topic, body, assignee)
 }
 
-func (t *Ticket) UUID() TicketUUID        { return t.uuid }
-func (t *Ticket) Ticketer() *Ticketer     { return t.ticketer }
-func (t *Ticket) Topic() *Topic           { return t.topic }
-func (t *Ticket) Body() string            { return t.body }
-func (t *Ticket) ExternalID() string      { return t.externalID }
-func (t *Ticket) SetExternalID(id string) { t.externalID = id }
-func (t *Ticket) Assignee() *User         { return t.assignee }
+func (t *Ticket) UUID() TicketUUID { return t.uuid }
+func (t *Ticket) Topic() *Topic    { return t.topic }
+func (t *Ticket) Body() string     { return t.body }
+func (t *Ticket) Assignee() *User  { return t.assignee }
 
 // Context returns the properties available in expressions
 //
@@ -68,26 +61,19 @@ func (t *Ticket) Context(env envs.Environment) map[string]types.XValue {
 //------------------------------------------------------------------------------------------
 
 type ticketEnvelope struct {
-	UUID       TicketUUID                `json:"uuid"                   validate:"required,uuid4"`
-	Ticketer   *assets.TicketerReference `json:"ticketer"               validate:"omitempty,dive"`
-	Topic      *assets.TopicReference    `json:"topic"                  validate:"omitempty,dive"`
-	Body       string                    `json:"body"`
-	ExternalID string                    `json:"external_id,omitempty"`
-	Assignee   *assets.UserReference     `json:"assignee,omitempty"     validate:"omitempty,dive"`
+	UUID     TicketUUID             `json:"uuid"                   validate:"required,uuid4"`
+	Topic    *assets.TopicReference `json:"topic"                  validate:"omitempty"`
+	Body     string                 `json:"body"`
+	Assignee *assets.UserReference  `json:"assignee,omitempty"     validate:"omitempty"`
 }
 
-// ReadTicket decodes a contact from the passed in JSON. If the ticketer or assigned user can't
+// ReadTicket decodes a contact from the passed in JSON. If the topic or assigned user can't
 // be found in the assets, we report the missing asset and return ticket without those.
 func ReadTicket(sa SessionAssets, data []byte, missing assets.MissingCallback) (*Ticket, error) {
 	e := &ticketEnvelope{}
 
 	if err := utils.UnmarshalAndValidate(data, e); err != nil {
 		return nil, err
-	}
-
-	ticketer := sa.Ticketers().Get(e.Ticketer.UUID)
-	if ticketer == nil {
-		missing(e.Ticketer, nil)
 	}
 
 	var topic *Topic
@@ -107,22 +93,15 @@ func ReadTicket(sa SessionAssets, data []byte, missing assets.MissingCallback) (
 	}
 
 	return &Ticket{
-		uuid:       e.UUID,
-		ticketer:   ticketer,
-		topic:      topic,
-		body:       e.Body,
-		externalID: e.ExternalID,
-		assignee:   assignee,
+		uuid:     e.UUID,
+		topic:    topic,
+		body:     e.Body,
+		assignee: assignee,
 	}, nil
 }
 
 // MarshalJSON marshals this ticket into JSON
 func (t *Ticket) MarshalJSON() ([]byte, error) {
-	var ticketerRef *assets.TicketerReference
-	if t.ticketer != nil {
-		ticketerRef = t.ticketer.Reference()
-	}
-
 	var topicRef *assets.TopicReference
 	if t.topic != nil {
 		topicRef = t.topic.Reference()
@@ -134,53 +113,9 @@ func (t *Ticket) MarshalJSON() ([]byte, error) {
 	}
 
 	return jsonx.Marshal(&ticketEnvelope{
-		UUID:       t.uuid,
-		Ticketer:   ticketerRef,
-		Topic:      topicRef,
-		Body:       t.body,
-		ExternalID: t.externalID,
-		Assignee:   assigneeRef,
+		UUID:     t.uuid,
+		Topic:    topicRef,
+		Body:     t.body,
+		Assignee: assigneeRef,
 	})
-}
-
-// Ticketer represents a ticket issuing system.
-type Ticketer struct {
-	assets.Ticketer
-}
-
-// NewTicketer returns a new classifier object from the given classifier asset
-func NewTicketer(asset assets.Ticketer) *Ticketer {
-	return &Ticketer{Ticketer: asset}
-}
-
-// Asset returns the underlying asset
-func (t *Ticketer) Asset() assets.Ticketer { return t.Ticketer }
-
-// Reference returns a reference to this classifier
-func (t *Ticketer) Reference() *assets.TicketerReference {
-	if t == nil {
-		return nil
-	}
-	return assets.NewTicketerReference(t.UUID(), t.Name())
-}
-
-// TicketerAssets provides access to all ticketer assets
-type TicketerAssets struct {
-	byUUID map[assets.TicketerUUID]*Ticketer
-}
-
-// NewTicketerAssets creates a new set of ticketer assets
-func NewTicketerAssets(ticketers []assets.Ticketer) *TicketerAssets {
-	s := &TicketerAssets{
-		byUUID: make(map[assets.TicketerUUID]*Ticketer, len(ticketers)),
-	}
-	for _, asset := range ticketers {
-		s.byUUID[asset.UUID()] = NewTicketer(asset)
-	}
-	return s
-}
-
-// Get returns the ticketer with the given UUID
-func (s *TicketerAssets) Get(uuid assets.TicketerUUID) *Ticketer {
-	return s.byUUID[uuid]
 }